Refactor capture component and events
